+++ scripts/size-stats/index.js
@@ -31,8 +31,8 @@ const buildApp = () => {
 
 const calcStats = () => {
     buildApp();
-    const appInitial = 146383; // precalculated - see webpack.config.js
-    const appInitialGzip = 42497;
+    const appInitial = 139844; // precalculated - see webpack.config.js
+    const appInitialGzip = 45086;
     const appWithMistica = getTotalSize(glob.sync(join(PATH_APP_BUILD, '**/*.js')));
     const appWithMisticaGzip = getTotalSize(glob.sync(join(PATH_APP_BUILD, '**/*.js')), {gzip: true});