+++ src/__stories__/header-story.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import {
     Header,
     HeaderLayout,
-    ButtonPrimary,
-    ButtonSecondary,
     Stack,
     NavigationBreadcrumbs,
     ResponsiveLayout,
-    Text3,
+    Placeholder,
+    Text1,
+    Text2,
 } from '..';
 import {useTextField, useCheckbox} from './helpers';
 
@@ -18,20 +18,11 @@ export default {
     },
 };
 
-const FieldWithCheckbox = ({children}: {children: React.ReactNode}) => (
-    <div style={{display: 'flex', alignItems: 'center'}}>{children}</div>
-);
-
 export const Default: StoryComponent = () => {
     const [pretitle, pretitleTextField] = useTextField('Pretitle', 'Your last bill');
     const [title, titleTextField] = useTextField('Title', 'December bill is now available');
-    const [preamount, preamountTextField] = useTextField('Preamount', 'Monthly fee (IVA included)');
-    const [amount, amountTextField] = useTextField('Amount', '60,44 €');
-    const [buttonLabel, buttonLabelTextField] = useTextField('Button', 'Download bill');
-    const [secondaryButtonLabel, secondaryButtonLabelTextField] = useTextField('SecondaryButton', 'Pay bill');
-    const [subtitle, subtitleTextField] = useTextField('Subtitle', 'This is a subtitle');
+    const [description, descriptionTextField] = useTextField('Description', 'This is a description');
     const [isInverse, inverseCheckbox] = useCheckbox('Inverse', true);
-    const [isErrorAmount, errorAmountCheckbox] = useCheckbox('Error amount', false);
     const [withExtraContent, extraContentCheckbox] = useCheckbox('With extra content', true);
     const [extraSideBySide, extraSideBySideCheckbox] = useCheckbox(
         'Extra content placed on the right in desktop',
@@ -56,31 +47,8 @@ export const Default: StoryComponent = () => {
                             />
                         ) : undefined
                     }
-                    header={
-                        <Header
-                            pretitle={pretitle}
-                            title={title}
-                            preamount={preamount}
-                            amount={amount}
-                            button={
-                                buttonLabel ? (
-                                    <ButtonPrimary href="asdf">{buttonLabel}</ButtonPrimary>
-                                ) : undefined
-                            }
-                            secondaryButton={
-                                secondaryButtonLabel ? (
-                                    <ButtonSecondary href="asdf">{secondaryButtonLabel}</ButtonSecondary>
-                                ) : undefined
-                            }
-                            subtitle={subtitle}
-                            isErrorAmount={isErrorAmount}
-                        />
-                    }
-                    extra={
-                        withExtraContent ? (
-                            <Text3 regular>some account chart here, for example</Text3>
-                        ) : undefined
-                    }
+                    header={<Header pretitle={pretitle} title={title} description={description} />}
+                    extra={withExtraContent ? <Placeholder /> : undefined}
                 />
             </div>
             <ResponsiveLayout>
@@ -88,16 +56,7 @@ export const Default: StoryComponent = () => {
                     {breadcrumbsCheckbox}
                     {pretitleTextField}
                     {titleTextField}
-                    {preamountTextField}
-                    <FieldWithCheckbox>
-                        {amountTextField}
-                        {errorAmountCheckbox}
-                    </FieldWithCheckbox>
-                    {buttonLabelTextField}
-                    {secondaryButtonLabelTextField}
-                    {subtitleTextField}
-                    {bleedCheckbox}
-                    {bleedValueTextField}
+                    {descriptionTextField}
                     {inverseCheckbox}
                     {extraContentCheckbox}
                     {extraSideBySideCheckbox}
@@ -108,6 +67,47 @@ export const Default: StoryComponent = () => {
 };
 
 Default.storyName = 'Header';
+
+/**
+ * The header is optional in order to allow webviews to delegate the header visualization to the surrounding native app.
+ * For example, in Novum App, the Start tab's greeting is rendered nativelly in the apps and via web in desktop.
+ */
+export const NoHeader: StoryComponent = () => {
+    const [extraSideBySide, extraSideBySideCheckbox] = useCheckbox(
+        'Extra content placed on the right in desktop',
+        true
+    );
+    return (
+        <Stack space={16}>
+            <div data-testid="header-layout">
+                <HeaderLayout
+                    isInverse
+                    sideBySideExtraOnDesktop={extraSideBySide}
+                    breadcrumbs={
+                        <NavigationBreadcrumbs
+                            title="Bills"
+                            breadcrumbs={[{title: 'Account', url: '/consumptions'}]}
+                        />
+                    }
+                    extra={<Placeholder />}
+                />
+            </div>
+            <ResponsiveLayout>
+                <Stack space={16}>
+                    <Text2 medium>Documentation:</Text2>
+                    <Text1 regular>
+                        The header is optional in order to allow webviews to delegate the header visualization
+                        to the surrounding native app. For example, in Novum App, the Start tab's greeting is
+                        rendered nativelly in the apps and via web in desktop.
+                    </Text1>
+                    {extraSideBySideCheckbox}
+                </Stack>
+            </ResponsiveLayout>
+        </Stack>
+    );
+};
+
+NoHeader.storyName = 'Header layout with no header';
 
 export const RichTexts: StoryComponent = () => {
     const filler = ' - more text'.repeat(20);
@@ -121,10 +121,6 @@ export const RichTexts: StoryComponent = () => {
                             truncate: true,
                         }}
                         title="Title is always a plain string"
-                        subtitle={{
-                            text: `Subtitle (truncated to two lines) ${filler}}`,
-                            truncate: 2,
-                        }}
                     />
                 }
             />