+++ src/skins/blau.tsx
@@ -63,9 +63,10 @@ export const getBlauSkin: GetKnownSkin = () => {
             background: palette.white,
             backgroundContainer: palette.white,
             backgroundBrand: palette.blauBluePrimary,
+            backgroundBrandSecondary: palette.blauBluePrimary,
             backgroundOverlay: applyAlpha(palette.blauBlueSecondary, 0.75),
             backgroundSkeleton: palette.grey2,
-            backgroundSkeletonInverse: applyAlpha(palette.white, 0.2),
+            backgroundSkeletonInverse: applyAlpha(palette.white, 0.3),
             navigationBarBackground: palette.blauBluePrimary,
             backgroundAlternative: palette.blauBluePrimary20,
             backgroundFeedbackBottom: palette.blauBluePrimary,
@@ -147,7 +148,7 @@ export const getBlauSkin: GetKnownSkin = () => {
             textPrimary: palette.grey6,
             textPrimaryInverse: palette.white,
             textSecondary: palette.grey5,
-            textSecondaryInverse: palette.white,
+            textSecondaryInverse: palette.blauBluePrimary20,
 
             // STATES
             error: palette.blauRed,
@@ -185,6 +186,7 @@ export const getBlauSkin: GetKnownSkin = () => {
         darkModeColors: {
             // BACKGROUNDS
             backgroundBrand: palette.darkModeBlack,
+            backgroundBrandSecondary: palette.darkModeBlack,
             background: palette.darkModeBlack,
             appBarBackground: palette.darkModeGrey,
             backgroundContainer: palette.darkModeGrey,