+++ src/skins/o2.tsx
@@ -69,6 +69,7 @@ export const getO2Skin: GetKnownSkin = () => {
             background: palette.white,
             backgroundContainer: palette.white,
             backgroundBrand: palette.o2BluePrimary,
+            backgroundBrandSecondary: palette.o2BluePrimary,
             backgroundOverlay: applyAlpha(palette.grey6, 0.6),
             backgroundSkeleton: palette.grey2,
             backgroundSkeletonInverse: applyAlpha(palette.white, 0.2),
@@ -152,7 +153,7 @@ export const getO2Skin: GetKnownSkin = () => {
             textPrimary: palette.grey6,
             textPrimaryInverse: palette.white,
             textSecondary: palette.grey5,
-            textSecondaryInverse: palette.white,
+            textSecondaryInverse: palette.o2BluePrimary15,
 
             // STATES
             error: palette.pepper,
@@ -190,6 +191,7 @@ export const getO2Skin: GetKnownSkin = () => {
             background: palette.darkModeBlack,
             backgroundContainer: palette.darkModeGrey,
             backgroundBrand: palette.darkModeBlack,
+            backgroundBrandSecondary: palette.darkModeBlack,
             backgroundOverlay: applyAlpha(palette.darkModeGrey, 0.8),
             backgroundSkeleton: palette.darkModeGrey6,
             backgroundSkeletonInverse: palette.darkModeGrey6,