+++ src/skins/types.tsx
@@ -7,7 +7,7 @@ export type GetKnownSkin = (variant?: SkinVariant) => KnownSkin;
 
 export type FontWeight = 'light' | 'regular' | 'medium' | 'bold';
 
-type TextPresetName = 'text5' | 'text6' | 'text7' | 'text8' | 'text9' | 'text10';
+type TextPresetName = 'text5' | 'text6' | 'text7' | 'text8' | 'text9' | 'text10' | 'cardTitle';
 export type TextPresetsConfig = {
     [preset in TextPresetName]: {
         weight: FontWeight;
@@ -40,6 +40,7 @@ export type Colors = {
     background: string;
     backgroundContainer: string;
     backgroundBrand: string;
+    backgroundBrandSecondary: string;
     backgroundOverlay: string;
     backgroundSkeleton: string;
     backgroundSkeletonInverse: string;